pyproject.toml
[project]
name = "rag-project-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "beautifulsoup4>=4.14.2",
    "docling>=2.57.0",
    "faiss-cpu>=1.12.0",
    "ipykernel>=7.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-docling>=1.1.0",
    "langchain-google-genai>=2.1.12",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.10",
    "pydantic>=2.12.2",
    "pymupdf4llm>=0.0.27",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "wikipedia>=1.4.0",
]
